Refactored, added tests for updating combinations/new words #97
